EDIT abstract out jws verification into JwsUtil
With the addition of the ReturnHandler, we now have two places where JWT
signature verification needs to be done. Instead of copying the same
code, I've tried to abstract it out into a public method in JwsUtil.

The JWT Framework provides an easy and a hard way to interact with it.
The hard way allows a lot more control over the process. I've found it
necessary to read the payload of a JWT before signature verification,
which required using part of the "hard" way. JwsUtil could've continued
using the hard way for signature verification, but I found there was a
lot of boilerplate that I just didn't want to bother with. So I've
decided to just reserialize back into a token string to be fed back into
the "easy" way.

Sadly, I think this has made what is an already convoluted process even
harder to follow, even though the code is simpler. But hopefully, this
makes JWT related security checks easier to manage in the future. E.g.:
the newer security standard is adding more supported algoritms for
signatures, this should make it easy to adapt.