FIX jwt typ/kid misplaced, exp not allowing leeway
This started from trying to abstract out the expiration time (1 hour)
into a shared constant in Param.

Turns out the JWT expiration time checking didn't allow for a 1 minute
leeway (in case of clock skew). I did intend to allow for the leeway in
checking but seemed to have undone it with a bit of puzzling code in
JwsUtil. I think the line in question ($now += self::TOKEN_LEEWAY;) was
probably added for testing purposes and then forgotten.

I think the old age token check was partially covering this up, because
the TOKEN_OLD_AGE was incorrectly set to the same as the expiration time
of 1 hour. So I've set the TOKEN_OLD_AGE to be 10 times the expiration
time.

The typ and kid fields in the return token in AuthRespHandler was being
duplicated. I was setting them in the JWT header when creating the
token, but they were also set as regular claims as part of the payload.
I've removed the duplicate ones in the claim payload.

Tests were also fixed.