EDIT deep link ReturnHandler actual implementation
Similar to the AuthRespHandler, we receive a JWT, validate it, and then
relay it. Except in the opposite direction. We receive a JWT from the
target tool and need to send a JWT back to the originating platform.

This should be a POST request with only 1 param named 'JWT'.

This is expected to take place in a call to the deep link return url
originally established by the DeepLinkFilter.

The JWT this time isn't as complicated as launch. The main payload is
the content items, which I don't think we can filter effectively, so is
just being passed through as is for now.

Like the launch, there is a state field that was saved in the DeepLink
table entry and needs to be relayed back. Note that deep link calls this
field "data". I've decided to refer to this as 'state' since 'data' was
too generic.

There is a set of message and logging params (normal and error) that
I've mirrored into the LtiLog. These messages are also passed through as
is.

The LTI version number '1.3.0' is now set as a constant in Param.