ADD DeepLink ReturnTest to test ReturnHandler
Test deep link's return stage. This is the last stage, where we receive
a return from the target tool and have to build our own return to send
to the originating platform.

Since this also involves a lot of JWT validation, this is very similar
to the AuthRespHandler. With less parameters and claims to check, I can
see a few improvements to how AuthRespHandler is laid out. But I'm not
sure if they can be reflected back. Instead of having a parameter for
only adding additional claims when creating the token, I've decided to
use that parameter to add every claim. This helps in testing that we
properly return an error when a required claim is missing.

The key difference with JWT validation from AuthResp is that AuthResp is
Platform to Tool while Return is reversed, Tool to Platform. So the keys
and various identifiers also needs to be reversed, case in point is
'iss' and 'aud' are reversed from what they were in AuthResp.

I also just realized that we didn't really test if nonces are
automatically marked as used once seen, so I've added a test to that
effect.

Also realized that we didn't really test if the formUrl being used to
send our responses weren't being tested, so I've added a check for it in
AuthRespTest and in ReturnTest.