EDIT lti launch now calls deep link implementation
Removed the routes that called on the original launch implementation.
The launch call should now go through the deep link implementation.

There is no longer separate routes for the platform/tool side for the
launch, since each handler will perform both sides at the same time. So
the route has been simplified to:
* /lti/launch/login
* /lti/launch/auth
* /lti/launch/redirect

Note the the last redirect route. This is the the auth response step.
The step before it is the auth request. It seems that it's more common
to refer to the auth response step as the "redirect" step, so I've
changed the route naming to match.

While the target_link_uri is set to midway, we don't actually direct the
user to the midway handler yet.