FIX login_hint not being passed back to platform
login_hint was sent in the OIDC login from the originating platform, and
we've stored it in the LtiSession so we can send it back in the auth
request to the originating platform.

However, the AuthReqHandler forgot to send login_hint back. So Canvas
failed the request because it couldn't find the user session.

This was fixed and the test modified to check for login_hint.