EDIT hook up deep link launch to midway
The deep link launch should now function identically to the original
launch code, since it properly directs users to the midway page before
continuing on to the tool.

Midway used to be split up between arrival and departure, but with this
change, we can get rid of departure and make it a single step. The route
for midway can also be correspondingly shortened from
'/lti/launch/midway/arrival' to '/lti/launch/midway'.

The new MIDWAY_REDIRECT_URI param tells us the target tool's endpoint
for the auth resp. This is the one that gets submitted to when the user
clicks the submit button.

The new MIDWAY_SESSION param tells us the LtiSession for this launch.

The params that used to be submitted was just the LTI_MESSAGE_HINT, this
is now changed to id_token and state.

I noticed that deep link launches were also stopping at midway for
instructors. There's no reason for the midway lookup to show up in a
deep link workflow, since the instructor is just selecting resources to
link to in the LMS. So I've added an exception for deep link requests.

Previously, when you arrive a midway, the name was appropriate because
it was actually the middle handoff between the tool and platform sides
of the shim. However, with the deep link implementation, there is
actually only one last auth resp that needs to be submitted to the
target tool. So the name is probably not really suitable anymore.