FIX deep link does not require resource link claim
The RESOURCE_LINK_URI claim is not required to be present in deep link
claims. I've added special treatment if the message type indicates a
deep link request to not check for its presence in deep link launches.

I've added it to ID_TOKEN_OPTIONAL_CLAIMS since it doesn't exactly
belong in ID_TOKEN_REQUIRED_CLAIMS anymore. While it's not really a
good fit, since it is required in other message types, I don't know a
better solution currently.