FIX deep link return not storing nonce
Since the nonce in the deep link return JWT wasn't generated by the
shim, we don't know about it and it gets wrongly treated as an invalid
nonce.

We need to add the nonce into the nonce store before checking validity.
If the nonce is already in the store, no harm done, since the store just
ignores the add.

The tests were hiding this incorrect behavior because I was using
Nonce::create() to generate the default nonce, which automatically adds
the nonce to the nonce cache. I've switched it to a hardcoded constant
instead, simulating that this nonce was generated by someone else.

Note that this has uncovered some problematic nonce expiry handling.
Namely that nonce expiry window should probably be as large as the
TOKEN_OLD_AGE window. Reason being that if the nonce expiry is smaller
than TOKEN_OLD_AGE, then once the nonce expired it is considered valid
again, and a replay attack is possible if sent before TOKEN_OLD_AGE.

This means that testRejectExpiredNonce() makes no sense. Since the
expired nonce should actually be valid again. I was wrongly treating
nonce expiry in the same manner as JWT expiry.