FIX SeedLtiInfo fails if no shim platform/tool
The whole reason that SeedLtiInfo exists is to seed in the shim's own
platform/tool entries. But it was failing on
getOwnPlatform()/getOwnTool() because it didn't catch the exception
being thrown by them.

Also, the shim's tool's client_id wasn't set to the proper config value,
which would've caused issues later on.