UPDATE composer & npm packages
Note that npm reports 37 moderate severity vulnerabilities involving
postcss. However, they're in the dependencies where we seem to be
up-to-date, so not much we can do. Guessing it's a new vulnerability in
postcss that hasn't been widely addressed yet, so we just need to wait
for the libraries to update.