EDIT webpack.mix.js add cache busting
Encountered an issue where the staging server was using cached js. Turns
out we can make Laravel Mix add a hash to filenames to prevent this from
happening.

Made it so that we only generate the hash for filenames when compiling
for production and only generate source maps when not in prod.

Also upgraded the node version in Dockerfile as it was complaining that
the npm version included in node:14 might have issues with the lock
file. Apparently the lock file went from v1 to v2 somewhere along the
way when I upgraded npm in Laradock.