EDIT launch login route pass target tool in path
The target_tool_id was being passed as a query, e.g.
(/login?target_tool_id=1). This might cause issues if platforms doesn't
correctly handle url rewrites, e.g. if they do it naively, just
appending to the end of the url string, then we have an invalid login
url. So it's better to pass it as part of the path, e.g.
(/login/tool/1).

Note that the shim tool entry's oidc_login_url is no longer accurate,
since we now have per-tool login url. So I've put in 'toolId' as the
placeholder param for the oidc_login_uri in the seeded shim tool entry.

Also updated getShimTargetLinkUriAttribute() since it's no longer
per-tool and should just go to midway. We're going to keep using the
attribute even though it can just be a constant now just in case we need
to change it to a per-tool thing somewhere down the line.