EDIT tools table add enable_midway_lookup boolean
This allows us to turn off the midway lookup tool for the entire tool.
This is in case the midway lookup causes issues during launch for the
entire tool. With the newer Ltijs update, the Ltijs only allows a 10
second validity for the id_token return. This means that if you spent
any time on the lookup tool, Ltijs will reject your LTI launch as
expired.

Added a MidwayTest to make sure that the behaviors around the midway
lookup is correct.