EDIT lti_sessions add 'state' json column
The 'token' JSON column was originally only used to store the id_token
from the auth resp step. However, it got co-opted to also store all
sorts of state used during the launch. This results in a unnecessarily
complex side effect where the previously stored state all gets removed
when the id_token comes in.

The easiest solution is to just add a 'state' column for storing the
state between the steps and only use the 'token' column for storing the
claims in id_token.