FIX possible replay attack against midway
Just realized that midway is vulnerable against replay attacks. This is
because the nonce that is part of the LTI launch is checked by the tool
and midway access happens before this tool check. So there is
essentially no nonce protection on access to midway.

The easiest solution is to add nonce protection LtiSession's encrypted
ID, which is just a JWE. The JWT spec says we should use the jti field
for this propose, so that's what I used.