ADD LaunchDirector to allow single side launches
The irony is that we just got rid of single side launches.

Currently, the shim acts as a relay for launches, playing both the
platform and tool role. Turns out there is a need for only acting as a
tool or only as a platform. E.g.: When stopping at the midway lookup, if
you stop for too long before continuing on to the target tool, the
target tool might consider your launch to be expired. In such cases, we
could use the existing LtiSession data and do a shim launch to the target
tool (without getting the originating platform involved), to get a
"fresh" launch session.

Luckily, the handlers were very easy to split apart into pretty much
independent send/recv functions. This should allow us to mix and match
these send/recv calls to do one sided launches. The new LaunchDirector
is meant to handle this mixing and matching. This commit only ports the
existing functionality over, so it's only a start.

Another scenario is if we want a launch that's only meant for the midway
lookup tool. In this case, the shim only needs to act as a tool and
doesn't need to contact the target tool at all.