ADD midway lookup tool only launch
This launch is meant just for accessing the midway lookup tool. Since we
now have a feature to disable midway entirely for a tool, we need
another way of getting to the midway lookup tool. I initially thought we
could just add a different login route, but this required configuring
and entirely new developer key entry in Canvas and as a result generated
a new client ID. To reworked the shim to allow multiple client IDs per
platform-tool pair would be too much work, so that idea was abandoned.

To do a midway lookup tool only launch, you need to configure the
'target_link_uri' param that is send during login to add a '/lookup' to
the end of the path. E.g.: the normal value is something like
'http://localhost/lti/launch/midway' and this will do a regular launch.
If you change it to 'http://localhost/lti/launch/midway/lookup', this
will do a midway lookup tool only launch. It'll bring you to the lookup
tool without the button allowing you to continue on to the target tool.

target_link_uri was chosen because this seems to be easily modifiable in
Canvas. When you add an "External Tool" link, the "URL" field is passed
on as the target_link_uri.

Since we only need to access the shim, we can skip talking to the target
tool entirely and do an entire launch where we only need to call the
shim's tool side.

I initally thought about adding a new database column to LtiSession to
mark a session as lookup only, however, the new state column made much
more sense. So there is now a 'sessionType' param (defined in
Param::SESSION_TYPE) that tells us what kind of session this is. The
idea is that the LaunchDirector can check the session to see how to
handle the launch. I initally thought the LaunchDirector would be only
one that needs to be aware, but it turns out the handlers needed to know
more. There's two types of session type right now: regular
(Param::SESSION_TYPE_REG) and midwayLookupOnly
(Param::SESSION_TYPE_LOOKUP).

Basic sanity check tests were added to check that midway lookup tool
requests work.