ADD quick and dirty Piazza custom params filter
Piazza requires a few custom parameters (user_id, logon_id, course_id).
Wrote a quick filter to replace the user_id with the LtiFakeUser id, the
logon_id with the LtiFakeUser sub, and the course_id with the
CourseContext id.

Note that this commit allows the custom claim to go through. I'm going
to have to think up a more flexible custom claim filter, since we
shouldn't have to write a custom filter for every tool.