REMOVE bootstrap & associated deps
Bootstrap required jquery and popper.js, both of which were also
removed. Sass and sass-loader were also removed, the plan is to use
PostCSS from now on.

We were using Bootstrap 4, which is now outdated since Bootstrap 5 is
out.

It looks like Laravel is moving to TailwindCSS, which was what I wanted
to use originally but decided not to since I didn't want to deviate from
Laravel defaults. Now that both Bootstrap and Laravel has moved on from
the old defaults, I think we can give integrating TailwindCSS a try.