FIX midway platform lookup not showing fake user
Not sure if this was always broken or if there was some backend Laravel
change that broke this. Basically, when you're looking up platform user
identities, you'll only see the user's real info. The fake user is not
displayed. This is because the query's fake user name gets overridden by
the real user name (in the database, the name column has the same name,
so when php converts it to an associative array, they're using the same
'name' key and overrides each other). The same happens with the
student_name column.

Explicitly giving each lti_real_user name & student_number column an
alias in the select seems to fix this.