FIX admin ui multiple notifications on login error
When your session is expired, the interceptor for unauthorized error
fires multiple times. Since the error notification was sent in the
interceptor, this means we get multiple error messages too. Moved the
error notification into the authStore's logout action, since that's only
called once.

When you used the wrong username or password, the login will fail, but
gets treated as a success. This results in us redirecting to the admin
page resulting in multiple errors due to the previous issue. Turns out
this is due to me forgetting to return the axios call to the login url,
which broke the promise chain prematurely. As the previous call in the
chain was the call to get the csrf token, which was a success, the
entire login was resolved as a success too. Returning the axios call
fixes the promise chain which then resolves the login failure properly.

Now that a failed login doesn't do a redirect, we also need to re-enable
the login button on error since it was disabled when the user first
clicked it.