EDIT admin sidebar, ind pg for platform/tool/usr
The previous admin page was just a long list of all the
platforms/tools/users. This is now separated out so that they all have
individual pages.

This also meant that we need a new navigation for these pages. This was
added as a sidebar on desktops which collapses into a topbar on mobile.
This new nav is the AdminNav component.  The previous topbar nav was
removed from all signed in pages and should only be visible on the
splash & login page. This is now named the Banner component.

The relative path imports were breaking when I moved the PlatformAdmin
components into the views directory to become PlatformView. So I added a
webpack alias '@admin' that should resolve to admin UI's js root.

I'm envisioning PlatformClient as a part of the Platform page, that's
why it hasn't been integrated yet.