EDIT admin api PlatformClient now a nested route
Instead of /api/platform-client/{id}, the api route is now
/api/platform/{platform id}/client/{platform client id}

This is to help with the admin UI change where we want to move the
platform client listing to be under the platform they're configured for.

Found out that the dependency injection in the controllers do not like
it when there are parameters missing. On routes where both the Platform
and PlatformClient is implicitly in the route, I can't just put
PlatformClient in the function definition. It'll want to pass the
Platform id as a string and cause an error cause the type declarations
doesn't match.

Added a simple PlatformClientControllerTest, since we should start
testing the APIs too.