EDIT admin ui platform integrates platform client
Renamed "platform client" as "allowed tools". Hopefully, this makes more
intuitive sense for users as we're essentially adding tools that can
access the given platform.

This has been moved so that each platform has its own list of allowed
tools.

I'd forgotten about the mobile first principle of tailwind, which
assumes that the base style is for mobile and you override for larger
screen sizes. This meant we had to rework some of the css which assumed
that we were basing on the desktop view and override for mobile.

Removed the hover highlight on the plainTable as it was broken by the
addition of the allowed tools td in the platform list.

Also realized that the vuex store abstraction moduleBase does not work
for nested resources like platform client. I've just made a platform
client specific implementation for now to get us working. But if we have
other nested resources in the future, we should figure out how to make
an abstracted nested resource vuex module.

There was an delete client action in the platformStore that doesn't seem
to be used, so I've deleted it.

The breaking out the back button on clicking "Edit" for a platform is
annoying, we should probably fix this.