EDIT admin ui platform pg more back btn compatible
The "Add Platform" and "Edit" platform buttons counter-intuitively does
not work with the back button. This is because they're not "real"
navigation, just toggling visibility.

I've added them to vue-router so they are real navigation, making them
work with the back button. They all still use the same PlatformView
component though, so we had to set a unique router-view key so the
components will probably reload on navigation.

This has the interesting side effect of removing the need for
catching/listening to emit events for these two actions.

Haven't figured out a nice way to do the same with the Allowed Tools
section.