EDIT admin ui tool & user pgs back btn compat
Made the tools & users pages also back button compatible. Add/Edit are
now child routes.

Realized that the Save & Cancel buttons on the add/edit forms, which was
originally just toggling visibility of the form, isn't consistent with
the page navigation. So I've replaced it with a navigation call to their
respective admin pages. Note that the form components are still emitting
an event for this, it should be only the *View components that are tied
to the router.