EDIT admin ui ToolForm enable_midway_lookup field
The enable_midway_lookup field sets whether instructor launches to the
tool should include a midway stop where they can lookup student
identities. Some tools might not work well with this, due to session
expiration times, so we should have the option of disabling this.

If the midway lookup is not possible, a dedicated lookup launch can be
done by appending a /lookup to the OIDC login url.