EDIT admin ui restore logged in account editing
The logged in user used to be able to edit their own account, setting a
new password. This broke when we moved to entirely vue routing since
that was dependent on a blade template. This commit restores it.

A new item was added to the nav menu for "Account". I've added icons for
the Account and Logout items so they can shrink down to just the icons
on mobile.

This required a new API path for getting information on the currently
logged in user. This was previously being passed in by the blade
template. Added a UserControllerTest for this new API route. We'll
slowly increase test coverage of the admin API this way. Interestingly,
the apiResource route for user kept overriding the new /api/user/self
call, wrongly treating the 'self' as an user id and calling
UserController's show(). Moving the self route declaration above the
apiResource one fixes this. But a test was added to make sure those two
api routes aren't conflicting with each other.