ADD fake_email_domain config, uuid fake email gen
Fake email addresses are generated based on a name based UUID computed
from the fake user's sub & name. The idea is to allow us to regenerate a
fake user's email based on only info that's available to the tool.

The new lti.fake_email_domain config tells us what domain to use for the
fake email addresses. This is set to env var FAKE_EMAIL_DOMAIN if set,
if not, then it falls back to "example.com".

A basic LtiFakeUser model test was added to check on email generation.