EDIT admin ui moduleBase avoid dup api calls
I noticed that PlatformClientAdmin was sending out multiple duplicate
API calls to get the list of tools. So decided to try to prevent this by
modifying moduleBase.

Only implemented for get & getAll actions. This will store ongoing
promises in state.waitingFor. At the start of the getAll & get actions,
it'll check to see if there is already an ongoing promise stored in
state.waitingFor, and if there is, that promise is returned instead of
us creating a new one. When the promise is resolved, it'll remove itself
from state.waitingFor.

As a side bug, I noticed that PlatformClientAdmin sometimes throw an
exception because it tries to access the tools variable before tools are
actually loaded. So I added a check to make sure that the tool is
present before rendering that section.

Also added a spinner that is visible on inital load to indicate that
PlatformClientAdmin is waiting for data (e.g. waiting for tools or
platformClient data to arrive back from the api).