EDIT admin ui rewrite instructions how to use shim
Given that the LTI endpoints have underwent a rewrite, the original
ShimConfigInfo was severely outdated. It was also a bit confusing. So
I've rewritten it into a step-by-step instruction instead.

The LTI dictionary is kind of an after thought, so it could probably be
improved. I figure it's probably good to define the terms used by the
step-by-step guide first.

The selected platform and tool in ShimConfigInfo is persisted in the
root vuex store. This is so that user selections aren't lost when they
navigate away, such as when they go to tool/platform pages for add
tool/platform. I figure it doesn't need an entire dedicated vuex store
so have just shoved it in the root store.

The API call to get the right endpoints were also rewritten so that we
have to give the tool and platform ids. This technically isn't necessary
for the API, since it doesn't need to know the platform, but I figure
it's more consistent this way. While the API doesn't need the platform,
the UI needs it to be able to load the correct list of PlatformClients
(aka: Allowed Tools).