ADD anonymization col to LtiFakeUser & related api
Added two two columns to lti_fake_users:

* is_anonymized - the user has chosen to use an anonymized identity,
  defaults to true
* enable_first_time_setup - tells us that this is a newly created fake
  identity and we need to show them the first time setup. First time
  setup right now is basically just setting the is_anonymized field

A related api call at /api/midway/config/anonymization/{fakeUserId}
was also created to set the is_anonymized value. This will also
set enable_first_time_setup to false.

A simple test case for the MidwayApiController was also added to test
these functionalities.

The AuthRouteAPI method in the User model was removed. I'm confused on
why it was there, as a precursory test shows that nothing was broken
when I removed it. And a quick search shows that there seems to be no
user reason for it. Making a note here in case I find out it breaks
something.