FIX JwksTest sometimes fail due to return order
The keys returned by the JWKS api is not sorted, so can return keys in
different orders. However, assertJson() cares about order and so we
sometimes get spurious test failures that would go away when we just
rerun the test.

The consensus seems to be that, if you want to avoid this, convert the
json to array and use phpunit's array assertions. In particular
assertEqualsCanonicalizing() assertion will reorder both input arrays
with the same sort function before checking the content. So this is what
I've used.

Due to the random nature of the failures, I can't be sure that this is
really fixed yet. Indeed, my last fix for this issue seemed to have
worked for a while but I've recently been running into this failure a
lot. It doesn't seem to happen much (if at all) when you run just the
JwksTest by itself. However, if you run it as part of the entire test
suite, the failure rate seems to increase. Let's hope it really is fixed
this time.