EDIT midway move continue form to own component
The continuation form submits the last step of the LTI launch to the
target tool. Since we're going to have another midway view now that also
needs the continuation form, I'm abstracting it out to its own
component.

Not sure why I was using a template to pass in the continuation form
inputs. Probably a leftover from when I was using blade templates more.
I've moved the id_token and state params into props so they're treated
same as all the other params we're passing to vue.

The form submit method doesn't have to be dynamic. So I've removed it as
a parameter and just going to use post.

Also added a @midway alias to webpack so we can use it to point to
midway ui's root js for imports.