REPLACE Jose\Easy\Build::JWS with JwsToken::build
With this commit, the only use left of Jose\Easy is in the tests. Not
going to migrate the tests for now due to time constraints. Moved
Jose\Easy to dev dependencies since only the tests need it.

JwsToken::build() ended up just straight passing through to
JwsUtil::build(). I'd prefer JwsToken being refactored in the future to
handle bother build and load more smoothly, but leaving it for now.

The 'typ' param is apparently supposed to go into the JWT header. It
seems that all our JWS uses always sets 'typ' to 'jwt', so I just
hardcoded that into JwsUtil::build().