ADD deep link content item filtering
When a deep link content item is of type "ltiResourceLink", it provides
a launch (OIDC login) URL that might be different from the normal launch
URL. So we need to store this URL for use when the shim launches into
the target tool.

The deep_link_content_items table has been added to store this url.
Initially thought about just modifying the existing deep_links table to
store this url, but it looks like each deep link request can return
multiple content items, each potentially with a different url in the
worse case.

Specs/DeepLink/ now has its first filter. The ContentItemFilter will run
on each DeepLink return, making sure that the originating platform sees
a shim url for each content item. I've made ContentItemFilter logging to
default to log level instead of the debug level in older filters as I'd
like to make the shim logging more verboose in general.

The shim's content item launch url is just the normal launch url with a
content item section added in:
    /launch/login/tool/{toolId}/dlci/{deepLinkContentItemId}

This lets us retrieve the content item which will tell us what url to
use when the shim acts like a platform to launch into the tool.

Test for the ContentItemFilter was added to the deep link ReturnTest. A
new test was added for content item launch named ContentItemTest.

LTI link content items actually has a list of custom params that you are
supposed to passback to the tool on launch. I initially thought that
we'd have to store this, but then realized that since Canvas must
have already implementing this book keeping, we can let Canvas do all
the work and just pass it on. This turns out to be a relief because it
looks like Turnitin deep link content items uses a param substitution
feature that would be quite impossible for us to implement.