FIX tool/platform edit fail missing params in keys
When editing existing tool/platform, the save fails because some params
in the elements inside the keys array are missing. This turned out to be
a Laravel validator change, where if you're validating array keys, the
validator will leave out any array keys that you haven't specified. This
led to the validator dropping crucial fields such as key id. Fixed by
adding in all the other array keys so the validator knows about them.

This fix was also required for create tool/platform, though with the
addition of only 1 field.

Also added imports to Laravel logging facilities to some classes.