FIX can't set anonymization option, missing csrf
For first time users, when they reach midway and are asked whether they
want to use an anonymized identity, they are unable to save their
selection anonymization option. This is due to a missing CSRF token.

Not sure when, but probably one of the Laravel updates must have made
CSRF requirements stricter. So I've added an explicit CSRF request to
the midway javascript.