EDIT first time setup show fake/real identity info
First time launch where user is prompted to pick whether they want to
use an anonymized identity, show the generated fake identity so the user
isn't caught offguard when they launch into the tool as what seems to be
the wrong user.

Right now also displaying real identity, but maybe we can omit that part
and only show the fake info.

Reworded the "is hosted outside of Canada" part in the intro paragraph
as that's no longer necessarily always true.